fix: Resolve BLE integration compilation issues
- Fix BleConnectionState enum usage throughout codebase
  - Replace sealed class patterns with enum patterns
  - Update when statements for proper enum handling
  - Add missing isConnected() extension function imports
- Clean up duplicate and conflicting BLE classes
  - Remove problematic MrdSdkDataSource implementation
  - Consolidate MRD SDK integration in single adapter
  - Update dependency injection configuration
- Resolve import conflicts and missing references
  - Standardize BLE model imports from domain.repository
  - Fix method call mismatches in BleRepositoryImpl
  - Update MrdProtocolAdapter with local state classes
- Ensure stable compilation with proper error handling
  - All BLE-related files now compile successfully
  - Maintain backward compatibility with existing code

+++ app/src/main/java/com/wishring/app/ble/BleAutoConnectService.kt
@@ -357,7 +357,7 @@ class BleAutoConnectService : Service() {
         return NotificationCompat.Builder(this, BleConstants.NOTIFICATION_CHANNEL_ID)
             .setContentTitle(title)
             .setContentText(content)
-            .setSmallIcon(icon)
+            .setSmallIcon(icon as Int)
             .setPriority(NotificationCompat.PRIORITY_LOW)
             .setOngoing(true)
             .setAutoCancel(false)
